+++ src/GoDefinitionProvider.ts
@@ -2,6 +2,7 @@
 import * as vscode from 'vscode';
 import { ReferenceProvider } from './ReferenceProvider';
 import { FileData } from './ReferenceProvider';
+import fs = require('fs');
 
 export default class GoDefinitionProvider implements vscode.DefinitionProvider {
 
@@ -9,9 +10,6 @@ export default class GoDefinitionProvider implements vscode.DefinitionProvider {
 		document: vscode.TextDocument,
 		position: vscode.Position,
 		token: vscode.CancellationToken): Thenable<vscode.Location> {
-
-		// Load the ativated XML file and replace the element Id with id
-		var DOMParser = require('xmldom').DOMParser;
 
 		// Get the selected word
 		const word = ReferenceProvider.getSelectedWord(document, position);
@@ -33,8 +31,70 @@ export default class GoDefinitionProvider implements vscode.DefinitionProvider {
 				files.push(new FileData(doc.uri, doc.getText().replace(/( )(Id=|Id =|Id  =)/gi, " id=")))
 		}
 
+		// Run this code only if user open a directory workspace
+		if (vscode.workspace.rootPath) {
+
+			var promise = vscode.workspace.findFiles(new vscode.RelativePattern(vscode.workspace.rootPath as string, '{**/*.xml}'))
+				.then((uris) => {
+					uris.forEach((uri) => {
+
+						// Check if the file is open. If yes, take precedence over unsaved version
+						var openedFile = files.filter(x => x.Uri.fsPath == uri.fsPath)
+
+						if (openedFile == null || openedFile.length == 0) {
+							var data = fs.readFileSync(uri.fsPath, 'utf8');
+							files.push(new FileData(uri, data.toString().replace(/( )(Id=|Id =|Id  =)/gi, " id=")));
+						}
+					});
+				}).then(() => {
+					console.log("End of list");
+					return this.processSearch(word, document, files, position);
+				}
+				);
+
+			return promise;
+		}
+		else {
+			return this.processSearch(word, document, files, position);
+		}
+
+	}
+
+	private processSearch(
+		word: String,
+		document: vscode.TextDocument,
+		files: FileData[],
+		position: vscode.Position): Thenable<vscode.Location> {
+
+		// Load the ativated XML file and replace the element Id with id
+		var DOMParser = require('xmldom').DOMParser;
+		var hierarchyFiles: FileData[] = [];
+
+		for (var i = 0; i < files.length; i++) {
+			var xmlDoc = new DOMParser().parseFromString(files[i].Data.toLowerCase());
+			var trustFrameworkPolicyElement = xmlDoc.getElementsByTagName("TrustFrameworkPolicy".toLowerCase());
+
+			if (trustFrameworkPolicyElement.length == 1) {
+				files[i].Policy = trustFrameworkPolicyElement[0].getAttribute("policyid");
+			}
+
+			var basePolicyElement = xmlDoc.getElementsByTagName("BasePolicy".toLowerCase());
+			if (basePolicyElement.length == 1) {
+
+				var policyid = basePolicyElement[0].getElementsByTagName("policyid");
+				if (policyid.length == 1) {
+
+					files[i].ParentPolicy = policyid[0].textContent;
+				}
+			}
+
+			console.log(files[i].Uri)
+		}
+
+		hierarchyFiles = this.getHierarchy(files, files[0], hierarchyFiles, 1);
+
 		// Iterate through files array
-		for (const file of files) {
+		for (const file of hierarchyFiles) {
 
 			var xmlDoc = new DOMParser().parseFromString(file.Data.toLowerCase());
 
@@ -42,19 +102,42 @@ export default class GoDefinitionProvider implements vscode.DefinitionProvider {
 			var nsAttr = xmlDoc.getElementById(word.toLowerCase());
 
 			// If element found and it's not the same element the user pointing (same file and same line)
-			if (nsAttr != null && 
+			if (nsAttr != null &&
 				!(file.Uri === document.uri && (nsAttr.lineNumber == position.line || nsAttr.lineNumber - 1 == position.line))) {
-				
-					// Return the selected element
+
+				var location = new vscode.Location(
+					file.Uri,
+					new vscode.Position(nsAttr.lineNumber - 1, nsAttr.columnNumber));
+
+				// Return the selected element
 				return new Promise(resolve => {
-					resolve(new vscode.Location(
-						file.Uri,
-						new vscode.Position(nsAttr.lineNumber - 1, nsAttr.columnNumber)));;
+					resolve(location);;
 				});
 			}
 		}
 
+
 		// Return no found (null)
 		return new Promise((resolve) => resolve());
+	}
+
+	private getHierarchy(files: FileData[], file: FileData, hierarchyFiles: FileData[], level: number): FileData[] {
+
+		if (level == 1) {
+			file.Level = level;
+			hierarchyFiles.push(file);
+		}
+
+		var paretnPolicy: FileData[] = files.filter(x => x.Policy == file.ParentPolicy);
+
+		if (paretnPolicy != null && paretnPolicy.length > 0) {
+			level++;
+			paretnPolicy[0].Level = level;
+			hierarchyFiles.push(paretnPolicy[0]);
+
+			hierarchyFiles = this.getHierarchy(files, paretnPolicy[0], hierarchyFiles, level);
+		}
+
+		return hierarchyFiles.sort(x => x.Level).reverse();
 	}
 }