+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## 18 October 2018
 - Adding Azure Application Insights trace log integration. Learn more [here](https://github.com/yoelhor/aad-b2c-vs-code-extension/blob/master/src/help/app-insights.md)
+- Go definition - VS code extension searches the definitions in all files in the working directory 
+- Go definition hierarchical search - VS code extension searches the definitions only in the parents policies 
 
 ## 18 May 2018
 - **Go to definition** - Always take precedence of editor open files, over unsaved version from file system (workspace folder)